refactor commands into command group files
